+++ abra/inference/bayesian/delta.py
@@ -2,7 +2,8 @@
 # -*- coding: utf-8 -*-
 from abra.config import DEFAULT_ALPHA, DEFAULT_BAYESIAN_INFERENCE_METHOD
 from abra.inference.inference_base import InferenceProcedure
-from abra.stats import Samples, Traces
+from abra.stats import Samples
+from abra.vis import Traces
 from abra.inference.bayesian import get_stan_model, get_model_datatype, get_model_data
 from abra.inference.bayesian.results import BayesianTestResults
 import numpy as np
