+++ abra/inference/bayesian/results.py
@@ -11,7 +11,7 @@ class BayesianTestResults(HypothesisTestResults):
     ----------
     alpha : float in (0, 1)
         The "significance" level, or type I error rate for the experiment
-    traces : instance of stats.Traces
+    traces : instance of abra.vis.Traces
         The results of a Bayesian inference procedure.
     hdi : tuple
         The values and percentiles associated with the lower and upper bounds of
